Added cases button, ethics tab in workspace
Attempted to work on save functionality by adding a NodePositionController class, which will store all references to nodes and their positions to save to a file. Unfortunately, I can't find a common ground for the nodes. Further, I worked on the cases tab, and ethics tab. I used some code from online to make the tabs draggable (as in, you can drag them around the screen and detach them).